Improve organisation of test cases
